Fix GraphSketcher's own warnings under warnings-as-errors
With warnings-as-errors re-enabled, fix the three warnings in GraphSketcher's own code
that are not in the demoted (vendored-noise) categories:

- RSAxis-XML.m: cast RSAxisType to int for nameFromAxisType() (-Wshorten-64-to-32).
- OSStyle-AttributeExtensions.m: remove a function-local static that shadowed the
  file-scoped defaultTabStopStringCache (-Wshadow). The shadow was also a latent bug — the
  function populated the local, so the file-scoped cache read elsewhere was never set.
- NSFont-RSExtensions: remove the -fontWithSize: category method; modern NSFont provides it
  natively with the same descriptor-preserving behavior, and overriding a framework method
  from a category is undefined (-Wobjc-protocol-method-implementation).

Also advances the OmniGroup submodule (warnings-as-errors re-enabled + toolchain warning
demotions). The full Mac build is green again with warnings-as-errors ON.

+++ GraphSketcher/Model/XML/RSAxis-XML.m
@@ -253,7 +253,7 @@
     [xmlDoc setAttribute:@"min" double:[self min]];
     [xmlDoc setAttribute:@"max" double:[self max]];
     [xmlDoc setAttribute:@"userModifiedRange" string:stringFromBool([self userModifiedRange])];
-    [xmlDoc setAttribute:@"scale" string:nameFromAxisType(_axisType)];
+    [xmlDoc setAttribute:@"scale" string:nameFromAxisType((int)_axisType)];
     [xmlDoc setAttribute:@"width" real:(float)[self width]];
     if (_shape)
 	[xmlDoc setAttribute:@"end-shape" string:nameFromShape(_shape)];
